+++ ExpandWorldPrefabs/ExpandWorldPrefabs.cs
@@ -12,7 +12,7 @@ public class EWP : BaseUnityPlugin
 {
   public const string GUID = "expand_world_prefabs";
   public const string NAME = "Expand World Prefabs";
-  public const string VERSION = "1.11";
+  public const string VERSION = "1.12";
 #nullable disable
   public static Harmony Harmony;
 #nullable enable